Fix misc sound bugs (#845)
* Change play function to return current sound instead of undefined

Change play to also return sound if unable to play

* Change ADSR function to accept ratios of duration rather than absolute values

* Changed ADSR function to have a linear decay

The previous version of ADSR (Attack, decay, sustain, release) uses an exponential decay function. This may cause issues because the tail end of the decay does not line up with the next part of the sound, which can cause glitches. Thus a linear function is used instead, because the tail end of the decay will match up exactly. There is the added benefit of the function being easier to read and code.

* Change play to display error if a sound is already playing

* Fix error not displaying when calling unsafe play when sound is already playing

* Fix adsr issues
* Create absolute_adsr function that accepts durations in seconds
* Change instruments to use absolute_adsr
* Fix typo in adsr comments

* Remove absolute_adsr and tweak adsr envelope values for instruments

* Correctness fix for play function jsdoc